# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: KeyControl API
  version: "1.0.0"
  description: |
    # Overview
    Welcome to the official API documentation for the KeyControl gateway platform. This API provides the programmatic interface for securely managing, routing, and validating your API keys.

    ## Authentication
    This API is secured using Bearer authentication. With the exception of public authentication routes (`/api/auth/*`), all endpoints require a valid token passed in the `Authorization` header:

    `Authorization: Bearer <YOUR_TOKEN>`

    **Supported Token Types:**
    * **JWT:** Obtained via the standard user login flow.
    * **Master API Key:** A system-level key (prefixed with `mk-`) used for administrative access and automated integrations.
servers:
  - url: "{scheme}://{host}:{port}/api"
    description: Admin API
    variables:
      scheme:
        default: http
        enum: [http, https]
      host:
        default: localhost
      port:
        default: "3000"
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or Master API Key
      description: |
        Pass either:
        • A JWT returned by `POST /auth/login` (or `/auth/login/verify-2fa`).
        • The Master API Key (prefix `mk-`) generated via the Organization endpoints.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
            - METHOD_NOT_ALLOWED
            - PAYLOAD_TOO_LARGE
            - BAD_GATEWAY
            - GATEWAY_TIMEOUT
        details:
          description: Optional field-level validation details
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
    SuccessFlag:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    Endpoint:
      type: object
      required: [url_pattern, method]
      properties:
        id:
          type: integer
        url_pattern:
          type: string
        method:
          type: string
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        unique_path:
          type: string
        secret_api_key:
          type: string
        external_api_base_url:
          type: string
        description:
          type: string
          nullable: true
        timeout_seconds:
          type: integer
          nullable: true
        timeout_response_code:
          type: integer
          default: 504
        timeout_response_body:
          type: string
          default: '{"error": "Request timeout"}'
        timeout_response_type:
          type: string
          enum: [json, text, xml]
          default: json
        created_at:
          type: string
          format: date-time
        total_usage_count:
          type: integer
        endpoint_groups_count:
          type: integer
        last_used:
          type: string
          format: date-time
          nullable: true
    ResourceDetail:
      allOf:
        - $ref: "#/components/schemas/Resource"
        - type: object
          properties:
            endpoint_groups:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
                    nullable: true
                  resource_id:
                    type: integer
                  endpoints:
                    type: array
                    items:
                      $ref: "#/components/schemas/Endpoint"
...
paths:
...
  /resources/{id}:
    get:
      tags: [Resources]
      operationId: getResource
      summary: Get a resource with its endpoint groups and endpoints
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Resources]
      operationId: updateResource
      summary: Update a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, secret_api_key]
              properties:
                name:
                  type: string
                  minLength: 1
                secret_api_key:
                  type: string
                  minLength: 1
                external_api_base_url:
                  type: string
                external_api_url:
                  type: string
                description:
                  type: string
                  nullable: true
                timeout_seconds:
                  type: integer
                  nullable: true
                timeout_response_code:
                  type: integer
                  default: 504
                timeout_response_body:
                  type: string
                timeout_response_type:
                  type: string
                  enum: [json, text, xml]
                  default: json
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Resources]
      operationId: deleteResource
      summary: Delete a resource
      description: Also nullifies any request_logs referencing this resource.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessFlag"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Endpoint Groups ──────────────────────────────────────────────────
...